# mw-general-bff/apiAutogenerated/mw-general-bff/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  title: Masters way general-bff API
  version: "1.0"
servers:
- url: /general
paths:
...
  /plans:
    post:
      operationId: create-plan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.CreatePlanPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.PlanPopulatedResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.NoRightToChangeDayReportError'
          description: User doesn't have rights to create plan.
      summary: Create a new plan
      tags:
      - plan
      x-codegen-request-body-name: request
  /plans/{planId}:
    delete:
      operationId: delete-plan
      parameters:
      - description: plan UUID
        in: path
        name: planId
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.NoRightToChangeDayReportError'
          description: User doesn't have rights to delete plan.
      summary: Delete plan by UUID
      tags:
      - plan
    patch:
      operationId: update-plan
      parameters:
      - description: plan UUID
        in: path
        name: planId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.UpdatePlanPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.PlanPopulatedResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.NoRightToChangeDayReportError'
          description: User doesn't have rights to update plan.
      summary: Update plan by UUID
      tags:
      - plan
      x-codegen-request-body-name: request
...
  /users:
    get:
      description: Get users with pagination
      operationId: get-all-users
      parameters:
      - description: Page number for pagination
        in: query
        name: page
        schema:
          type: integer
      - description: Number of items per page
        in: query
        name: limit
        schema:
          type: integer
      - description: Part of user email for filters
        in: query
        name: email
        schema:
          type: string
      - description: Part of user name for filters
        in: query
        name: name
        schema:
          type: string
      - description: '''mentor'' | ''all'' status for filter'
        in: query
        name: mentorStatus
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.GetAllUsersResponse'
          description: OK
      summary: Get all users
      tags:
      - user
  /users/{userId}:
    get:
      operationId: get-user-by-uuid
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.UserPopulatedResponse'
          description: OK
      summary: Get user by UUID
      tags:
      - user
    patch:
      operationId: update-user
      parameters:
      - description: user ID
        in: path
        name: userId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/mw-general-bff_internal_schemas.UpdateUserPayload'
        description: query params
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/mw-general-bff_internal_schemas.UserPlainResponse'
          description: OK
      summary: Update user by UUID
      tags:
      - user
      x-codegen-request-body-name: request
...
components:
  schemas:
...
    mw-general-bff_internal_schemas.CreatePlanPayload:
      properties:
        dayReportUuid:
          type: string
        description:
          type: string
        isDone:
          type: boolean
        ownerUuid:
          type: string
        time:
          type: integer
      required:
      - dayReportUuid
      - description
      - isDone
      - ownerUuid
      - time
      type: object
...
    mw-general-bff_internal_schemas.GetAllUsersResponse:
      example:
        size: 0
        users:
        - createdAt: createdAt
          ownWays: 5
          mentoringWays: 1
          favoriteWays: 6
          favoriteForUsers: 0
          imageUrl: imageUrl
          name: name
          description: description
          isMentor: true
          uuid: uuid
          email: email
          tags:
          - name: name
            uuid: uuid
          - name: name
            uuid: uuid
        - createdAt: createdAt
          ownWays: 5
          mentoringWays: 1
          favoriteWays: 6
          favoriteForUsers: 0
          imageUrl: imageUrl
          name: name
          description: description
          isMentor: true
          uuid: uuid
          email: email
          tags:
          - name: name
            uuid: uuid
          - name: name
            uuid: uuid
      properties:
        size:
          type: integer
        users:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.UserPlainResponseWithInfo'
          type: array
      required:
      - size
      - users
      type: object
...
    mw-general-bff_internal_schemas.JobTagResponse:
      example:
        color: color
        name: name
        description: description
        uuid: uuid
      properties:
        color:
          type: string
        description:
          type: string
        name:
          type: string
        uuid:
          type: string
      required:
      - color
      - description
      - name
      - uuid
      type: object
...
    mw-general-bff_internal_schemas.NoRightToChangeDayReportError:
      example:
        message: message
      properties:
        message:
          type: string
      type: object
...
    mw-general-bff_internal_schemas.PlanPopulatedResponse:
      example:
        createdAt: createdAt
        ownerUuid: ownerUuid
        wayName: wayName
        ownerName: ownerName
        wayUuid: wayUuid
        description: description
        time: 6
        dayReportUuid: dayReportUuid
        isDone: true
        uuid: uuid
        tags:
        - color: color
          name: name
          description: description
          uuid: uuid
        - color: color
          name: name
          description: description
          uuid: uuid
        updatedAt: updatedAt
      properties:
        createdAt:
          type: string
        dayReportUuid:
          type: string
        description:
          type: string
        isDone:
          type: boolean
        ownerName:
          type: string
        ownerUuid:
          type: string
        tags:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.JobTagResponse'
          type: array
        time:
          type: integer
        updatedAt:
          type: string
        uuid:
          type: string
        wayName:
          type: string
        wayUuid:
          type: string
      required:
      - createdAt
      - dayReportUuid
      - description
      - isDone
      - ownerName
      - ownerUuid
      - tags
      - time
      - updatedAt
      - uuid
      - wayName
      - wayUuid
      type: object
...
    mw-general-bff_internal_schemas.UpdatePlanPayload:
      properties:
        description:
          type: string
        isDone:
          type: boolean
        time:
          type: integer
      type: object
...
    mw-general-bff_internal_schemas.UpdateUserPayload:
      properties:
        description:
          type: string
        imageUrl:
          type: string
        isMentor:
          type: boolean
        name:
          type: string
      type: object
...
    mw-general-bff_internal_schemas.UserPlainResponse:
      example:
        createdAt: createdAt
        imageUrl: imageUrl
        name: name
        description: description
        isMentor: true
        uuid: uuid
        email: email
      properties:
        createdAt:
          type: string
        description:
          type: string
        email:
          type: string
        imageUrl:
          type: string
        isMentor:
          type: boolean
        name:
          type: string
        uuid:
          type: string
      required:
      - createdAt
      - description
      - email
      - imageUrl
      - isMentor
      - name
      - uuid
      type: object
    mw-general-bff_internal_schemas.UserPlainResponseWithInfo:
      example:
        createdAt: createdAt
        ownWays: 5
        mentoringWays: 1
        favoriteWays: 6
        favoriteForUsers: 0
        imageUrl: imageUrl
        name: name
        description: description
        isMentor: true
        uuid: uuid
        email: email
        tags:
        - name: name
          uuid: uuid
        - name: name
          uuid: uuid
      properties:
        createdAt:
          type: string
        description:
          type: string
        email:
          type: string
        favoriteForUsers:
          type: integer
        favoriteWays:
          type: integer
        imageUrl:
          type: string
        isMentor:
          type: boolean
        mentoringWays:
          type: integer
        name:
          type: string
        ownWays:
          type: integer
        tags:
          items:
            $ref: '#/components/schemas/mw-general-bff_internal_schemas.UserTagResponse'
          type: array
        uuid:
          type: string
      required:
      - createdAt
      - description
      - email
      - favoriteForUsers
      - favoriteWays
      - imageUrl
      - isMentor
      - mentoringWays
      - name
      - ownWays
      - tags
      - uuid
      type: object
...
    mw-general-bff_internal_schemas.UserTagResponse:
      example:
        name: name
        uuid: uuid
      properties:
        name:
          type: string
        uuid:
          type: string
      required:
      - name
      - uuid
      type: object
...