# mw-general-bff/internal/controllers/plan.controller.go
package controllers

import (
	openapiGeneral "mw-general-bff/apiAutogenerated/general"
	"mw-general-bff/internal/auth"
	"mw-general-bff/internal/facades"
	"mw-general-bff/internal/schemas"
	"mw-general-bff/internal/services"
	"mw-general-bff/pkg/utils"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/samber/lo"
)

// Without next lines swagger does not see openapi models
var _ = &openapiGeneral.MwServerInternalSchemasCreatePlanPayload{}

type PlanController struct {
	planFacade *facades.PlanFacade
}

func NewPlanController(planFacade *facades.PlanFacade) *PlanController {
	return &PlanController{planFacade}
}

// Create Plan handler
// @Summary Create a new plan
// @Description
// @Tags plan
// @ID create-plan
// @Accept  json
// @Produce  json
// @Param request body schemas.CreatePlanPayload true "query params"
// @Success 200 {object} schemas.PlanPopulatedResponse
// @Failure 403 {object} schemas.NoRightToChangeDayReportError "User doesn't have rights to create plan."
// @Router /plans [post]
func (pc *PlanController) CreatePlan(ctx *gin.Context) {
	var payload *schemas.CreatePlanPayload

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	responseRaw, err := pc.planFacade.CreatePlan(ctx, payload)
	utils.HandleErrorGin(ctx, err)

	response := schemas.PlanPopulatedResponse{
		Uuid:          responseRaw.Uuid,
		CreatedAt:     responseRaw.CreatedAt,
		UpdatedAt:     responseRaw.UpdatedAt,
		Description:   responseRaw.Description,
		Time:          responseRaw.Time,
		OwnerUuid:     responseRaw.OwnerUuid,
		OwnerName:     responseRaw.OwnerName,
		IsDone:        responseRaw.IsDone,
		DayReportUuid: responseRaw.DayReportUuid,
		WayUUID:       responseRaw.WayUuid,
		WayName:       responseRaw.WayName,
		Tags: lo.Map(responseRaw.Tags, func(tag openapiGeneral.MwServerInternalSchemasJobTagResponse, _ int) schemas.JobTagResponse {
			return schemas.JobTagResponse{
				Uuid:        tag.Uuid,
				Name:        tag.Name,
				Description: tag.Description,
				Color:       tag.Color,
			}
		}),
	}

	ctx.JSON(http.StatusOK, response)
}

// Update Plan handler
// @Summary Update plan by UUID
// @Description
// @Tags plan
// @ID update-plan
// @Accept  json
// @Produce  json
// @Param request body schemas.UpdatePlanPayload true "query params"
// @Param planId path string true "plan UUID"
// @Success 200 {object} schemas.PlanPopulatedResponse
// @Failure 403 {object} schemas.NoRightToChangeDayReportError "User doesn't have rights to update plan."
// @Router /plans/{planId} [patch]
func (pc *PlanController) UpdatePlan(ctx *gin.Context) {
	var payload *schemas.UpdatePlanPayload
	planID := ctx.Param("planId")

	modifierUserIdRaw, _ := ctx.Get(auth.ContextKeyUserID)
	modifierUserId := modifierUserIdRaw.(string)

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"status": "Failed payload", "error": err.Error()})
		return
	}

	planRaw, err := pc.planFacade.UpdatePlan(ctx, &services.UpdatePlanParams{
		PlanID:           planID,
		Description:      payload.Description,
		Time:             payload.Time,
		IsDone:           payload.IsDone,
		ModifierUserUuid: modifierUserId,
	})
	utils.HandleErrorGin(ctx, err)

	plan := schemas.PlanPopulatedResponse{
		Uuid:          planRaw.Uuid,
		CreatedAt:     planRaw.CreatedAt,
		UpdatedAt:     planRaw.UpdatedAt,
		Description:   planRaw.Description,
		Time:          planRaw.Time,
		OwnerUuid:     planRaw.OwnerUuid,
		OwnerName:     planRaw.OwnerName,
		IsDone:        planRaw.IsDone,
		DayReportUuid: planRaw.DayReportUuid,
		WayUUID:       planRaw.WayUuid,
		WayName:       planRaw.WayName,
		Tags: lo.Map(planRaw.Tags, func(tag openapiGeneral.MwServerInternalSchemasJobTagResponse, _ int) schemas.JobTagResponse {
			return schemas.JobTagResponse{
				Uuid:        tag.Uuid,
				Name:        tag.Name,
				Description: tag.Description,
				Color:       tag.Color,
			}
		}),
	}

	ctx.JSON(http.StatusOK, plan)
}

// Delete Plan handler
// @Summary Delete plan by UUID
// @Description
// @Tags plan
// @ID delete-plan
// @Accept  json
// @Produce  json
// @Param planId path string true "plan UUID"
// @Success 204
// @Failure 403 {object} schemas.NoRightToChangeDayReportError "User doesn't have rights to delete plan."
// @Router /plans/{planId} [delete]
func (pc *PlanController) DeletePlanById(ctx *gin.Context) {
	planID := ctx.Param("planId")

	err := pc.planFacade.DeletePlanById(ctx, planID)
	utils.HandleErrorGin(ctx, err)

	ctx.Status(http.StatusNoContent)
}