# mw-general-bff/internal/controllers/user.controller.go
package controllers

import (
	openapiGeneral "mw-general-bff/apiAutogenerated/general"
	"mw-general-bff/internal/facades"
	"mw-general-bff/internal/schemas"
	"mw-general-bff/internal/services"
	"mw-general-bff/pkg/utils"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/samber/lo"
)

// Without next lines swagger does not see openapi models
var _ = &openapiGeneral.MwServerInternalSchemasUserPlainResponse{}

type UserController struct {
	toUserMentoringRequestFacade *facades.UserFacade
}

func NewUserController(toUserMentoringRequestFacade *facades.UserFacade) *UserController {
	return &UserController{toUserMentoringRequestFacade}
}

// @Summary Update user by UUID
// @Description
// @Tags user
// @ID update-user
// @Accept  json
// @Produce  json
// @Param request body schemas.UpdateUserPayload true "query params"
// @Param userId path string true "user ID"
// @Success 200 {object} schemas.UserPlainResponse
// @Router /users/{userId} [patch]
func (uc *UserController) UpdateUser(ctx *gin.Context) {
	var payload *schemas.UpdateUserPayload
	userID := ctx.Param("userId")

	if err := ctx.ShouldBindJSON(&payload); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	userRaw, err := uc.toUserMentoringRequestFacade.UpdateUser(ctx, &services.UpdateUserParams{
		UserID:      userID,
		Name:        payload.Name,
		Description: payload.Description,
		ImageUrl:    payload.ImageUrl,
		IsMentor:    payload.IsMentor,
	})
	utils.HandleErrorGin(ctx, err)

	user := schemas.UserPlainResponse{
		Uuid:        userRaw.Uuid,
		Name:        userRaw.Name,
		Email:       userRaw.Email,
		Description: userRaw.Description,
		CreatedAt:   userRaw.CreatedAt,
		ImageUrl:    userRaw.ImageUrl,
		IsMentor:    userRaw.IsMentor,
	}

	ctx.JSON(http.StatusOK, user)
}

// @Summary Get user by UUID
// @Description
// @Tags user
// @ID get-user-by-uuid
// @Accept  json
// @Produce  json
// @Param userId path string true "user ID"
// @Success 200 {object} schemas.UserPopulatedResponse
// @Router /users/{userId} [get]
func (uc *UserController) GetUserById(ctx *gin.Context) {
	userID := ctx.Param("userId")

	populatedUserRaw, err := uc.toUserMentoringRequestFacade.GetPopulatedUserById(ctx, uuid.MustParse(userID))
	utils.HandleErrorGin(ctx, err)

	customWayCollections := lo.Map(populatedUserRaw.CustomWayCollections, func(collection openapiGeneral.MwServerInternalSchemasWayCollectionPopulatedResponse, _ int) schemas.WayCollectionPopulatedResponse {
		return schemas.WayCollectionPopulatedResponse{
			Uuid: collection.Uuid,
			Name: collection.Name,
			Ways: lo.Map(collection.Ways, func(way openapiGeneral.MwServerInternalSchemasWayPlainResponse, _ int) schemas.WayPlainResponse {
				return schemas.WayPlainResponse{
					Uuid:            way.Uuid,
					Name:            way.Name,
					GoalDescription: way.GoalDescription,
					UpdatedAt:       way.UpdatedAt,
					CreatedAt:       way.CreatedAt,
					EstimationTime:  way.EstimationTime,
					IsCompleted:     way.IsCompleted,
					Owner: schemas.UserPlainResponse{
						Uuid:        way.Owner.Uuid,
						Name:        way.Owner.Name,
						Email:       way.Owner.Email,
						Description: way.Owner.Description,
						CreatedAt:   way.Owner.CreatedAt,
						ImageUrl:    way.Owner.ImageUrl,
						IsMentor:    way.Owner.IsMentor,
					},
					CopiedFromWayUuid: way.CopiedFromWayUuid.Get(),
					ProjectUuid:       way.ProjectUuid.Get(),
					IsPrivate:         way.IsPrivate,
					FavoriteForUsers:  way.FavoriteForUsers,
					DayReportsAmount:  way.DayReportsAmount,
					Mentors: lo.Map(way.Mentors, func(mentor openapiGeneral.MwServerInternalSchemasUserPlainResponse, _ int) schemas.UserPlainResponse {
						return schemas.UserPlainResponse{
							Uuid:        mentor.Uuid,
							Name:        mentor.Name,
							Email:       mentor.Email,
							Description: mentor.Description,
							CreatedAt:   mentor.CreatedAt,
							ImageUrl:    mentor.ImageUrl,
							IsMentor:    mentor.IsMentor,
						}
					}),
					MetricsDone:   way.MetricsDone,
					MetricsTotal:  way.MetricsTotal,
					ChildrenUuids: way.ChildrenUuids,
					WayTags: lo.Map(way.WayTags, func(tag openapiGeneral.MwServerInternalSchemasWayTagResponse, _ int) schemas.WayTagResponse {
						return schemas.WayTagResponse{
							Uuid: tag.Uuid,
							Name: tag.Name,
						}
					}),
				}
			}),
			CreatedAt: collection.CreatedAt,
			UpdatedAt: collection.UpdatedAt,
			OwnerUuid: collection.OwnerUuid,
			Type:      collection.Type,
		}
	})

	defaultWayCollections := schemas.DefaultWayCollections{
		Own: schemas.WayCollectionPopulatedResponse{
			Uuid:      populatedUserRaw.DefaultWayCollections.Own.Uuid,
			Name:      populatedUserRaw.DefaultWayCollections.Own.Name,
			CreatedAt: populatedUserRaw.DefaultWayCollections.Own.CreatedAt,
			UpdatedAt: populatedUserRaw.DefaultWayCollections.Own.UpdatedAt,
			OwnerUuid: populatedUserRaw.DefaultWayCollections.Own.OwnerUuid,
			Type:      populatedUserRaw.DefaultWayCollections.Own.Type,
			Ways:      lo.Map(populatedUserRaw.DefaultWayCollections.Own.Ways, ConvertWay),
		},
		Favorite: schemas.WayCollectionPopulatedResponse{
			Uuid:      populatedUserRaw.DefaultWayCollections.Favorite.Uuid,
			Name:      populatedUserRaw.DefaultWayCollections.Favorite.Name,
			CreatedAt: populatedUserRaw.DefaultWayCollections.Favorite.CreatedAt,
			UpdatedAt: populatedUserRaw.DefaultWayCollections.Favorite.UpdatedAt,
			OwnerUuid: populatedUserRaw.DefaultWayCollections.Favorite.OwnerUuid,
			Type:      populatedUserRaw.DefaultWayCollections.Favorite.Type,
			Ways:      lo.Map(populatedUserRaw.DefaultWayCollections.Favorite.Ways, ConvertWay),
		},
		Mentoring: schemas.WayCollectionPopulatedResponse{
			Uuid:      populatedUserRaw.DefaultWayCollections.Mentoring.Uuid,
			Name:      populatedUserRaw.DefaultWayCollections.Mentoring.Name,
			CreatedAt: populatedUserRaw.DefaultWayCollections.Mentoring.CreatedAt,
			UpdatedAt: populatedUserRaw.DefaultWayCollections.Mentoring.UpdatedAt,
			OwnerUuid: populatedUserRaw.DefaultWayCollections.Mentoring.OwnerUuid,
			Type:      populatedUserRaw.DefaultWayCollections.Mentoring.Type,
			Ways:      lo.Map(populatedUserRaw.DefaultWayCollections.Mentoring.Ways, ConvertWay),
		},
	}

	response := schemas.UserPopulatedResponse{
		Uuid:               populatedUserRaw.Uuid,
		Name:               populatedUserRaw.Name,
		Email:              populatedUserRaw.Email,
		Description:        populatedUserRaw.Description,
		CreatedAt:          populatedUserRaw.CreatedAt,
		ImageUrl:           populatedUserRaw.ImageUrl,
		IsMentor:           populatedUserRaw.IsMentor,
		WayCollections:     customWayCollections,
		DefaultCollections: defaultWayCollections,
		FavoriteForUsers:   populatedUserRaw.FavoriteForUsers,
		FavoriteUsers: lo.Map(populatedUserRaw.FavoriteUsers, func(user openapiGeneral.MwServerInternalSchemasUserPlainResponse, _ int) schemas.UserPlainResponse {
			return schemas.UserPlainResponse{
				Uuid:        user.Uuid,
				Name:        user.Name,
				Email:       user.Email,
				Description: user.Description,
				CreatedAt:   user.CreatedAt,
				ImageUrl:    user.ImageUrl,
				IsMentor:    user.IsMentor,
			}
		}),
		Tags: lo.Map(populatedUserRaw.Tags, func(tag openapiGeneral.MwServerInternalSchemasUserTagResponse, _ int) schemas.UserTagResponse {
			return schemas.UserTagResponse{
				Uuid: tag.Uuid,
				Name: tag.Name,
			}
		}),
		WayRequests: lo.Map(populatedUserRaw.WayRequests, ConvertWay),
		Projects: lo.Map(populatedUserRaw.Projects, func(project openapiGeneral.MwServerInternalSchemasProjectPlainResponse, _ int) schemas.ProjectPlainResponse {
			return schemas.ProjectPlainResponse{
				ID:        project.Id,
				Name:      project.Name,
				IsPrivate: project.IsPrivate,
				UserIDs:   project.UserIds,
			}
		}),
	}

	ctx.JSON(http.StatusOK, response)
}

// @Summary Get all users
// @Description Get users with pagination
// @Tags user
// @ID get-all-users
// @Accept  json
// @Produce  json
// @Param page query integer false "Page number for pagination"
// @Param limit query integer false "Number of items per page"
// @Param email query string false "Part of user email for filters"
// @Param name query string false "Part of user name for filters"
// @Param mentorStatus query string false "'mentor' | 'all' status for filter"
// @Success 200 {object} schemas.GetAllUsersResponse
// @Router /users [get]
func (uc *UserController) GetAllUsers(ctx *gin.Context) {
	page := ctx.DefaultQuery("page", "1")
	limit := ctx.DefaultQuery("limit", "10")
	email := ctx.DefaultQuery("email", "")
	name := ctx.DefaultQuery("name", "")
	// mentorStatus = "mentor" | "all"
	mentorStatus := ctx.DefaultQuery("mentorStatus", "all")

	reqPage, err := strconv.Atoi(page)
	utils.HandleErrorGin(ctx, err)
	reqLimit, err := strconv.Atoi(limit)
	utils.HandleErrorGin(ctx, err)

	responseRaw, err := uc.toUserMentoringRequestFacade.GetAllUsers(ctx, &services.GetAllUsersParams{
		MentorStatus: mentorStatus,
		UserName:     name,
		UserEmail:    email,
		Page:         reqPage,
		Limit:        reqLimit,
	})
	utils.HandleErrorGin(ctx, err)

	response := schemas.GetAllUsersResponse{
		Size: int64(responseRaw.Size),
		Users: lo.Map(responseRaw.Users, func(user openapiGeneral.MwServerInternalSchemasUserPlainResponseWithInfo, _ int) schemas.UserPlainResponseWithInfo {
			return schemas.UserPlainResponseWithInfo{
				Uuid:             user.Uuid,
				Name:             user.Name,
				Email:            user.Email,
				ImageUrl:         user.ImageUrl,
				IsMentor:         user.IsMentor,
				Description:      user.Description,
				CreatedAt:        user.CreatedAt,
				FavoriteForUsers: user.FavoriteForUsers,
				FavoriteWays:     user.FavoriteWays,
				MentoringWays:    user.MentoringWays,
				OwnWays:          user.OwnWays,
				Tags: lo.Map(user.Tags, func(tag openapiGeneral.MwServerInternalSchemasUserTagResponse, _ int) schemas.UserTagResponse {
					return schemas.UserTagResponse{
						Uuid: tag.Uuid,
						Name: tag.Name,
					}
				}),
			}
		}),
	}

	ctx.JSON(http.StatusOK, response)
}